+++ plugins/governance/changelog.md
@@ -1,1 +1,1 @@
-- chore: update framework version to 1.1.59
+- chore: upgrades core to v1.2.48 and framework to 1.1.60